refactor: ScrollTopListener의 클릭리스너를 내부에서 설정하도록 변경
