+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/competitionList/CompetitionFragment.kt
@@ -9,7 +9,7 @@ import com.emmsale.R
 import com.emmsale.data.model.Event
 import com.emmsale.databinding.FragmentCompetitionBinding
 import com.emmsale.presentation.base.BaseFragment
-import com.emmsale.presentation.common.ScrollUpVisibilityListener
+import com.emmsale.presentation.common.ScrollTopListener
 import com.emmsale.presentation.common.extension.getSerializableExtraCompat
 import com.emmsale.presentation.common.views.FilterTag
 import com.emmsale.presentation.common.views.filterChipOf
@@ -62,7 +62,6 @@ class CompetitionFragment : BaseFragment<FragmentCompetitionBinding>() {
         initViewModel()
         initEventRecyclerView()
         initEventFilterButtonClickListener()
-        initScrollTopFabClickListener()
     }
 
     private fun initViewModel() {
@@ -72,7 +71,7 @@ class CompetitionFragment : BaseFragment<FragmentCompetitionBinding>() {
     private fun initEventRecyclerView() {
         binding.rvEvents.adapter = eventAdapter
         binding.rvEvents.addOnScrollListener(
-            ScrollUpVisibilityListener(requireContext(), binding.fabScrollTop),
+            ScrollTopListener(targetView = binding.fabScrollTop),
         )
     }
 
@@ -134,10 +133,6 @@ class CompetitionFragment : BaseFragment<FragmentCompetitionBinding>() {
 
     private fun initEventFilterButtonClickListener() {
         binding.btnEventFilter.setOnClickListener { navigateToEventFilter() }
-    }
-
-    private fun initScrollTopFabClickListener() {
-        binding.fabScrollTop.setOnClickListener { binding.rvEvents.smoothScrollToPosition(0) }
     }
 
     private fun navigateToEventDetail(event: Event) {
