+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/conferenceList/ConferenceFragment.kt
@@ -9,7 +9,7 @@ import com.emmsale.R
 import com.emmsale.data.model.Event
 import com.emmsale.databinding.FragmentConferenceBinding
 import com.emmsale.presentation.base.BaseFragment
-import com.emmsale.presentation.common.ScrollUpVisibilityListener
+import com.emmsale.presentation.common.ScrollTopListener
 import com.emmsale.presentation.common.extension.getSerializableExtraCompat
 import com.emmsale.presentation.common.views.FilterTag
 import com.emmsale.presentation.common.views.filterChipOf
@@ -60,7 +60,6 @@ class ConferenceFragment : BaseFragment<FragmentConferenceBinding>() {
         initViewModel()
         initEventRecyclerView()
         initEventFilterButtonClickListener()
-        initScrollTopFabClickListener()
     }
 
     private fun initViewModel() {
@@ -70,7 +69,7 @@ class ConferenceFragment : BaseFragment<FragmentConferenceBinding>() {
     private fun initEventRecyclerView() {
         binding.rvEvents.adapter = eventAdapter
         binding.rvEvents.addOnScrollListener(
-            ScrollUpVisibilityListener(requireContext(), binding.fabScrollTop),
+            ScrollTopListener(targetView = binding.fabScrollTop),
         )
     }
 
@@ -132,10 +131,6 @@ class ConferenceFragment : BaseFragment<FragmentConferenceBinding>() {
 
     private fun initEventFilterButtonClickListener() {
         binding.btnEventFilter.setOnClickListener { navigateToEventFilter() }
-    }
-
-    private fun initScrollTopFabClickListener() {
-        binding.fabScrollTop.setOnClickListener { binding.rvEvents.smoothScrollToPosition(0) }
     }
 
     private fun navigateToEventDetail(event: Event) {
